add acp118 aggregator & move legacy aggregator to warp tests
